+++ include/error/error_types/parsing/definitions/func/err_def_func_required_type_unknown.hpp
@@ -7,9 +7,9 @@ class ErrDefFuncRequiredTypeUnknown : public BaseError {
     ErrDefFuncRequiredTypeUnknown(  //
         const ErrorType error_type, //
         const Hash &file_file,      //
-        const size_t line,          //
-        const size_t column,        //
-        const size_t length         //
+        const unsigned int line,    //
+        const unsigned int column,  //
+        const unsigned int length   //
         ) :
         BaseError(error_type, file_file, line, column, length) {}
